+++ pybricks/pupdevices.h
@@ -15,6 +15,7 @@
 extern const mp_obj_module_t pb_module_pupdevices;
 
 extern const mp_obj_type_t pb_type_pupdevices_ColorDistanceSensor;
+extern const mp_obj_type_t pb_type_pupdevices_ColorLightMatrix;
 extern const mp_obj_type_t pb_type_pupdevices_ColorSensor;
 extern const mp_obj_type_t pb_type_pupdevices_ForceSensor;
 extern const mp_obj_type_t pb_type_pupdevices_InfraredSensor;